// lib/models/hike.model.ts
export interface HikeModel{
    trailname: string;
    distance: number;
    ranking: number;
    location: string;
    description: string;

    // constructor(tn: string, dist: number, rank: number, loc: string, desc: string) {
    //     this.trailname = tn;
    //     this.distance = dist;
    //     this.ranking = rank;
    //     this.location = loc;
    //     this.description = desc;
    // }
}